Moved RhoType.scala to models project
