+++ casper/src/main/scala/coop/rchain/casper/rholang/BlockRandomSeed.scala
@@ -4,8 +4,8 @@ import coop.rchain.casper.protocol.BlockMessage
 import coop.rchain.crypto.PublicKey
 import coop.rchain.crypto.hash.Blake2b512Random
 import coop.rchain.models.Par
+import coop.rchain.models.rholang.RhoType.Name
 import coop.rchain.models.syntax._
-import coop.rchain.rholang.interpreter.RhoType.Name
 import coop.rchain.rspace.hashing.Blake2b256Hash
 import scodec.bits.ByteVector
 import scodec.codecs.{bytes, uint8, utf8, variableSizeBytes, vlong}
