+++ casper/src/main/scala/coop/rchain/casper/rholang/sysdeploys/CheckBalance.scala
@@ -5,7 +5,7 @@ import coop.rchain.casper.rholang.types.{SystemDeploy, SystemDeployUserError}
 import coop.rchain.crypto.PublicKey
 import coop.rchain.crypto.hash.Blake2b512Random
 import coop.rchain.models.NormalizerEnv.{Contains, ToEnvMap}
-import coop.rchain.rholang.interpreter.RhoType.{Extractor, RhoNumber}
+import coop.rchain.models.rholang.RhoType.{Extractor, RhoNumber}
 
 class CheckBalance(pk: PublicKey, rand: Blake2b512Random) extends SystemDeploy(rand) {
