+++ casper/src/test/scala/coop/rchain/casper/rholang/Resources.scala
@@ -3,15 +3,11 @@ package coop.rchain.casper.rholang
 import cats.Parallel
 import cats.effect.{Concurrent, ContextShift, Resource, Sync}
 import cats.syntax.all._
-import coop.rchain.models.syntax._
 import coop.rchain.casper.storage.RNodeKeyValueStoreManager.rnodeDbMapping
-import coop.rchain.crypto.hash.Blake2b512Random
 import coop.rchain.metrics
 import coop.rchain.metrics.{NoopSpan, Span}
 import coop.rchain.models.Par
 import coop.rchain.rholang.Resources.mkTempDir
-import coop.rchain.rholang.interpreter.RhoRuntime.RhoHistoryRepository
-import coop.rchain.rholang.interpreter.RhoType.Name
 import coop.rchain.rspace.syntax._
 import coop.rchain.shared.Log
 import coop.rchain.store.LmdbDirStoreManager.mb
