+++ node/src/test/scala/coop/rchain/node/revvaultexport/RhoTrieTraverserTest.scala
@@ -1,21 +1,17 @@
 package coop.rchain.node.revvaultexport
 
 import cats.effect.Concurrent
-import com.google.protobuf.ByteString
 import coop.rchain.casper.genesis.contracts.{Registry, StandardDeploys}
 import coop.rchain.casper.helper.TestNode.Effect
 import coop.rchain.casper.helper.TestRhoRuntime.rhoRuntimeEff
 import coop.rchain.casper.syntax._
-import coop.rchain.crypto.hash.Blake2b512Random
 import coop.rchain.casper.util.{ConstructDeploy, GenesisBuilder}
+import coop.rchain.crypto.hash.Blake2b512Random
 import coop.rchain.metrics.{Metrics, NoopSpan, Span}
-import coop.rchain.models.GUnforgeable.UnfInstance.GPrivateBody
-import coop.rchain.models.{GPrivate, GUnforgeable, Par}
-import coop.rchain.rspace.hashing.Blake2b256Hash
+import coop.rchain.models.rholang.RhoType.Name
+import coop.rchain.models.syntax._
 import coop.rchain.shared.Log
 import monix.eval.Task
-import coop.rchain.models.syntax._
-import coop.rchain.rholang.interpreter.RhoType.Name
 import monix.execution.Scheduler.Implicits.global
 import org.scalatest.flatspec.AnyFlatSpec
